Avoid generating a sub-module for macro-declared items
Previously, the dyn-dyn procedural macros would generate their items in
a newly defined sub-module of the current scope with an implicit import
of super::* to import types from the parent scope. However, this could
cause issues if any type provided by the user used super itself or if
the type the macro was called upon was declared in a function.

To avoid these issues, these procedural macros now generate their
internal items in the current scope and simply prepend __dyn_dyn_ to
hopefully avoid naming conflicts.