Allow blanket generic impls of a generic dyn-dyn base trait
Previously, the design of how the trait bounds for dyn-dyn base traits
were created prevented blanket impls of generic dyn-dyn base traits
where a generic argument to the base trait was unrelated to the type
being implemented. In particular, the generic argument to the DynDynBase
trait was always a marker type retrieved from DynDynImpl::BaseMarker.
This would result in errors regarding unconstrained generics if an
attempt were made to add such a trait implementation.

This has been fixed by switching the generic argument to DynDynBase (now
renamed to DynDynDerived) to use the dyn object directly instead of a
marker type. This was previously not done since doing so directly
results in a compiler error due to a cycle while computing
super-predicates. To avoid this issue, a new marker trait with a blanket
impl for those types implementing DynDynDerived needs to be introduced.