Add test for implementing base trait on foreign type
Previously, it was not possible to implement a dyn-dyn base trait on a
foreign type, even if the base trait were defined in the current crate.
This was due to the fact that the compiler would complain that the
marker type was not local to the current crate, since the marker type
was retrieved via a trait from dyn-dyn.

As a consequence of the switch away from such marker types to allow
blanket impls, this now also works as expected. To ensure this continues
to work, a new test has been added for it.