Implement support for fat dyn-dyn pointers
Previously, it was necessary to dynamically call get_dyn_dyn_table from
the vtable every time a trait lookup was performed. This may not be
desirable from a performance perspective, since this function should be
pure and so the same value should be returned every time.

In order to get around this, a new DynDynFat struct has been added. This
struct wraps a pointer to a dyn-dyn base trait and caches the table so
that it does not have to be retrieved again.