Switch to using a macro for downcasts
Previously, a method on the DynDyn trait was used to actually perform a
downcast. While this integrated nicely with Rust's type system, it had
one particular limitation: it didn't play nice with auto marker traits.
It is sometimes necessary to propagate an extra marker bound like Send
or Sync from the base trait type to the derived trait type. Doing this
with generics requires jumping through a lot of hoops, including needing
to store extra unnecessary TypeIds. It also permanently restricted
handling of such markers to those hard-coded into dyn-dyn.

To alleviate this issue, a new macro dyn_dyn_cast! has been added to
perform the downcast instead. This macro allows code to be injected that
verifies the extra bounds statically, avoiding the runtime overhead of
the extra TypeIds and allowing all marker traits to be handled instead
of just Send and Sync.