Fix cargo test trying to run code from README.md
Apparently, by including README.md as the crate's doc comment, cargo
test was trying to compile and run the examples in it as doctests. This
is an issue since these examples are missing use declarations.
Additionally, even if these declarations were added, it seems that
proc-macro-crate has a bug that prevents the procedural macros from
working correctly.