Change the DynDyn trait to avoid a blanket Deref impl
Previously, the DynDyn trait had a blanket impl for all types that
dereference to a type implementing the base trait. This was nice, but it
was preventing a proper impl for DynDynFat from being added without the
very buggy specialization feature.

In order to work around this, the DynDyn trait no longer has a blanket
implementation for Deref. Instead, the dyn_dyn_cast! macro now handles
auto-deref internally using its method resolution-based specialization.
This avoids all the pitfalls of the current system and allows for a
proper DynDyn implementation for DynDynFat.