Remove dependency on proc-macro-crate
Previously, the proc-macro-crate crate was being used to allow the
procedural macros from dyn-dyn-macros to be used in unit tests in the
dyn-dyn crate. However, proc-macro-crate currently has an issue where it
does not work correctly in doctests, meaning that all examples in the
documentation had to be marked as untested.

Now that there are no longer any unit tests using these macros in
dyn-dyn, it is safe to switch to always using "::dyn_dyn" as the crate
name and remove the proc-macro-crate dependency.