Fix some issues with trait object lifetimes
It turns out that the current implementations of some of the inner
workings of the #[dyn_dyn_base] attribute and dyn_dyn_cast! had issues
dealing with trait objects with lifetimes attached. In particular, if
the trait object could not be proven to be 'static, then there were two
issues:

  - DynDynBase would not be properly implemented for the base trait
    object, since the impl generated by #[dyn_dyn_base] had an implicit
    'static lifetime bound
  - The dyn_dyn_cast! macro would complain that the reference was
    escaping the function since __dyn_dyn_deref_typecheck was defined to
    return a fake 'static reference to the underlying type