Fix unsoundness with dyn_dyn_cast! returned lifetime
Previously, dyn_dyn_cast! did not properly constrain the lifetime bound
of its result trait object. This would allow incorrect lifetime casting
from dyn Base + 'a to dyn Trait + 'b where 'a: 'b did not hold, e.g. if
'b was 'static.

This has been corrected by adding a useless if-else with an always-true
condition to allow the other side of this control flow to be used to
force the lifetime of the result to be constrained to be equal to the
lifetime of the input.