Fix dyn_dyn_cast! not keeping temporaries live
The previous implementation of the dyn_dyn_cast! macro was not properly
keeping any temporaries used in the computation of the value to be
downcast live until the end of the current statement. This would require
the caller to add let declarations to ensure these temporaries live long
enough, which was quite annoying.

This has now been corrected by ensuring that the input from the user is
used as the input to the closure without ever being put into a let
declaration. This did also necessitate some changes to how the
resolution-based specialization worked, since rustc has trouble
performing type-checking in this case unless a fully-qualified method
name is used to complete the dereference.