Rework the DynDyn trait to allow downcasting of smart pointers
Previously, the way the DynDyn trait worked only ever allowed
downcasting into &D or &mut D types. However, when downcasted types need
to be stored in a data structure, it might be beneficial to e.g.
downcast a Box<dyn Base> into a Box<dyn Trait>.

In order to accomodate this use case, the DynDyn trait has been
completely reworked, along with reworking how the dyn_dyn_cast! macro is
implemented. Now, when a dyn_dyn_cast! invocation starts with the
keyword `move`, it is possible to perform downcasts on smart pointers in
this manner.