Change dyn_dyn_cast! to return a Result
Previously, the dyn_dyn_cast! macro would return an Option that contains
the downcasted value if the downcast was valid and None otherwise.
However, this was suboptimal since there was no way to get the
originally passed in value if the downcast failed. This wasn't a big
deal back when only references could be used, but now that smart
pointers can be downcast, this has become an issue since the smart
pointer will be unconditionally dropped if the downcast isn't valid.

In order to combat this, dyn_dyn_cast! has been changed to return a
Result instead of an Option. This means that the original value can be
returned in an Err if the downcast fails.