Switch DynDynTableEntry::new to use Unsize
Previously, creating a DynDynTableEntry relied on passing in a closure
that performed a pointer downcast. This was a relic of the time before
dyn-dyn properly made use of the Rust compiler's `unsize` feature.
Because of this, creating a DynDynTableEntry had to be unsafe.

This has now been reworked to use the Unsize trait to perform the
downcast automatically. This eliminates the possibility of passing in a
closure that does not return a pointer with the correct metadata, making
it possible to safely call DynDynTableEntry::new.