Implement DynTrait for DST struct types
Previously, the DynTrait trait was only implemented if the type used in
its DynMetadata matched the outer type. This is true for trait objects
themselves, but is not true for DST structs, where the DynMetadata type
matches the inner trait object type.

While proper support for DST structs is still a ways off, the DynTrait
trait is being updated to allow DST structs in order to avoid the need
to change too much down the line in order to allow for DST struct
support.