Rename DynTrait to DynDynCastTarget
Previously, the DynTrait trait was used to represent all trait objects
that could be cast to. While the name made sense at the time, it makes
much less sense now that it is also implemented for DST structs
containing a trait object. It has now been renamed to reflect its actual
purpose: as a potential downcast target for a downcast.