Move AnyDynMetadata to the table module
Previously, AnyDynMetadata was contained in the cast_target module.
While this made sense back when it was heavily used by DynTrait (now
DynDynCastTarget), this no longer makes sense now that typed DynMetadata
is used there instead.

Since the table implementation is the only part of the crate that now
uses this struct, it seems to make much more sense to move it to the
table module.