Export some types that are part of the public API
Previously, the DynDynBase and DynDynDowncastTarget traits, as well as
the AnyDynMetadata struct, were not exported publicly for users to use.
This was strange, since these types are prominently used in several
public APIs.

In the case of DynDynDowncastTarget and AnyDynMetadata, these types were
declared in private modules and never re-exported. Because of this, it
was impossible for user code to even name them, which limited how
several public APIs could be used.

In the case of the DynDynBase trait, this trait *could* technically be
accessed by external code by going through the undocumented `internal`
module (so that the procedural macros can name it), but had no
documentation and no stable way of referring to it.

In order to make using dyn-dyn easier, these types are now exported
publicly in the root of the crate so that user code can name them.