Fix bad compiler error messages for dyn_dyn_cast! input type
When DerefHelper was rewritten when improving the DynDyn trait, a
significant regression occurred: if the input did not match any of the
specialization patterns, the compiler would complain that DerefHelper
did not implement DerefHelperEnd. The previous implementation did not do
this and would instead properly point out that the input type did not
implement DynDyn.

The cause of this was that the previous version had a fake version of
the "end" method that could never be called, since whenever its trait
bounds were met a different type would end up being used, but the new
version did not replicate this. In order to get the good error messages
back, DerefHelper now has a useless impl for DerefHelperEnd that can
never actually be used.