Allow dyn_dyn_cast! to be used with wrapped DST types
Previously, the dyn_dyn_cast! macro always expected that its input had
the base type `dyn Base`. However, when casting DST structs we actually
want to cast with the base type `Struct<dyn Base>`.

In order to support this use-case, a new syntax has been added to
dyn_dyn_cast! that allows the trait object type to be wrapped in a
generic type when performing the cast.

This technically now allows use of dyn-dyn with DST structs, but
requires that the DST struct manually implements the unsafe DynDynBase
trait.

Issue: #3