Reorganize code to put dyn-dyn at the root of the repo
Previously, the repo was organized so that the dyn-dyn and
dyn-dyn-macros crates were in subdirectories in the repo. However, this
doesn't really make much sense, since dyn-dyn-macros is really just an
implementation detail of dyn-dyn.

In order to better organize the code, the dyn-dyn crate now lives at the
root of the repo and the dyn-dyn-macros crate now lives in the macros
subdirectory.