Add #![forbid(unsafe_op_in_unsafe_fn)]
This change enables a lint against code that uses unsafe operations in
an unsafe fn without a surrounding unsafe block. This makes it more
clear where the difference between an API that is unsafe to call and the
unsafe operations it performs are. It also means that clippy now
complains if an unsafe operation inside an unsafe fn doesn't have a
safety comment.