Add SAFETY comments in the procedural macros
While all code in dyn-dyn had safety comments, this was not being
enforced in the code generated by the macros in dyn-dyn-macros. The
unsafe code generated by dyn_dyn_cast! in particular is nontrivial and a
lot of extra code that does nothing is actually generated to ensure that
it's safe, so it makes sense to document why it's actually safe.