Improve macro hygiene of dyn_dyn_impl
Previously, the dyn_dyn_impl procedural macro was not particularly
hygienic as it would define a new struct in the calling scope to hold
the DynDynTable internals. The macro tried to work around this and avoid
name conflicts by incorporating the caller's line number into the name
of this struct, but there was a fatal flaw with this approach: it does
not work when the call to dyn_dyn_impl was itself generated by a macro.

Since this generated struct is only actually used inside one method in
the generated code, there's a fairly easy answer to this problem: just
define the struct inside the method itself. This makes it impossible for
two of these structs to conflict with each other, since only one is ever
defined in a particular scope.