Remove UncheckedDowncast reliance on DynDynTarget
Previously, the UncheckedDowncast::downcast_unchecked method required
that the type being downcast to implemented DynDynTarget (i.e. that it
was provably a trait object). This was needlessly restrictive for what
this trait actually does and caused it to be difficult to actually
implement in some cases because the compiler sometimes has difficulties
proving that a type's metadata will be DynMetadata<_> if that type
involves generic arguments.

To address that, just allow any Pointee type to be used as the target
type in that method. The compiler has a much easier time dealing with
this and will not generally emit spurious errors due to it being unable
to typecheck an implementation.