Remove the B generic parameter from DowncastUnchecked
The previous implementation of DowncastUnchecked had a generic parameter
B meant to pass the base trait. However, nothing about this trait
actually requires access to the base trait and this was just a remnant
of a previous implementation that was scrapped before it saw the light
of day.

To simplify implementing this trait, this generic parameter has been
completely removed.