Remove reliance on the `effects' Nightly feature
Previously, DynDynCastTarget was marked with #[const_trait] so that it
could be used in the construction of a DynDynTableEntry. However, the
Nightly feature that makes this possible is being significantly reworked
in newer versions of Rust. Because of this, its usage here no longer
actually works.

In order to remedy this, the DynDynTableEntry::new method now directly
depends on an implementation of the Pointee trait and will no longer use
DynDynCastTarget. This allows it to still work in a const context
without needing to rely on #[const_trait].