Remove ability to use marker traits in dyn_dyn_cast!
This feature relied on the fact that for a trait Trait, the vtables for
Trait and Trait + Marker would always be equivalent. However, the
upcoming arbitrary_self_types feature will break this invariant,
meaning that it is no longer sound to take a vtable pointer from a
pointer cast to *const dyn Trait and use its vtable to construct a
*const dyn Trait + Marker.

While it's possible that this functionality could be brought back in a
more limited sense at some point in the future, e.g. by requiring users
to explicitly list which auto traits they want to be able to use in
dyn_dyn_cast! inside the #[dyn_dyn_impl(...)], it's clear that the
current version cannot remain since it will become unsound.